Refactor import statements in UserMapper and UserRegistrationMapper to use DTO package
